% Generated by roxygen2 (4.0.1): do not edit by hand
\name{countlines}
\alias{countlines}
\title{Count the number of lines in a file}
\usage{
countlines(fn)
}
\arguments{
\item{string}{input filepath}
}
\value{
numeric a integer for the number of lines
}
\description{
Count the number of lines in a file
}
\examples{
countlines(system.file("example/test.assoc.linear", package="txtutils")
}

