% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fileExists}
\alias{fileExists}
\title{Check if a file exists, if it does, do nothing, otherwise throw an exception}
\usage{
fileExists(name)
}
\arguments{
\item{string}{filepath}
}
\description{
Check if a file exists, if it does, do nothing, otherwise throw an exception
}
\examples{
fileExists(system.file("example/test.assoc.linear", package="txtutils")
}

