% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ncols}
\alias{ncols}
\title{Counts the number of columns of whitespace delimited file.}
\usage{
ncols(fn)
}
\arguments{
\item{string}{input filepath}
}
\value{
numeric a integer for the number of columns
}
\description{
Counts the number of columns of whitespace delimited file.
}
\examples{
ncols(system.file("example/test.assoc.linear", package="txtutils")
}

