% Generated by roxygen2 (4.0.1): do not edit by hand
\name{printlines}
\alias{printlines}
\title{print all the lines of a text file.}
\usage{
printlines(fn)
}
\arguments{
\item{string}{input filepath}
}
\description{
print all the lines of a text file.
}
\examples{
printlines(system.file("example/test.assoc.linear", package="txtutils")
}

