% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readcol}
\alias{readcol}
\title{Read one column of a whitespace delimited text file}
\usage{
readcol(fileName, colNum, nSkip, maxRowNum)
}
\arguments{
\item{string}{input filepath}

\item{integer}{the target column number}

\item{integer}{skip the first N lines}

\item{integer}{maximum number of lines to read}
}
\value{
vector a vector of strings containing the target column
}
\description{
Read one column of a whitespace delimited text file
}
\examples{
readcol(system.file("example/test.assoc.linear", package="txtutils"), 2, 1, 15)
}

