% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readcols}
\alias{readcols}
\title{Read columns of a whitespace delimited text file}
\usage{
readcols(fn, colsel, nFirstSkipLines, nSkipUnit)
}
\arguments{
\item{string}{input filepath}

\item{vector}{a vector of target column numbers}

\item{integer}{skip the first N lines}

\item{integer}{read one line out of every M}
}
\value{
matrix a matrix of strings from selected columns
}
\description{
Read columns of a whitespace delimited text file
}
\examples{
readcol(system.file("example/test.assoc.linear", package="txtutils"), 2, 1, 15)
}

