\name{txtutils-package}
\alias{txtutils-package}
\alias{txtutils}
\docType{package}
\title{
An auxiliary package for processing text files
}
\description{
An auxiliary package for processing text files
}
\details{
\tabular{ll}{
Package: \tab txtutils\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-08-13\cr
License: \tab GPL \cr
}
An auxiliary package for processing text files
}
\author{
Kaiyin Zhong

Maintainer: Kaiyin Zhong <kindlychung@gmail.com>
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
